Address Jiri's comments
